requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pyjwt==2.8.0
cryptography==42.0.2
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.3.1
redis==5.0.1
sentry_sdk==1.40.4
